requirements.txt
opencv-python-headless
numpy
pypotrace
svgwrite
scour
requests
boto3
streamlit
scipy
scikit-image
cairosvg
svgpathtools
matplotlib
Pillow
pytest
pytest-cov
pytest-mock
pytest-timeout
psutil